arch: add RV32 support with dual-XLEN assembly and build system
Add MM=32 build target:
- Makefile.compiler: conditional -march=rv32gc/ilp32d
- Makefile: -melf32lriscv linker emulation for RV32
- New include/arch/riscv/arch.h: XLEN/REGSZ macros
- swtch.S/trap_entry.S: sd/ld ↔ sw/lw via #if XLEN==64
- trap.c: mcause shift uses (XLEN-1) instead of hardcoded 63
- vmm.c: skip Sv39 on RV32 (#if __riscv_xlen==32 guard)
- fdt.h/fdt.c: unsigned long → unsigned long long for fdt_be64

Both architectures build cleanly with make (RV64) and make MM=32.
KV32 boots on qemu-system-riscv32 with all subsystems functional.

+++ include/fdt.h
@@ -32,10 +32,10 @@ static inline unsigned int fdt_be32(const void *p) {
   return ((unsigned int)b[0] << 24) | ((unsigned int)b[1] << 16) | ((unsigned int)b[2] << 8) | b[3];
 }
 
-static inline unsigned long fdt_be64(const void *p) {
+static inline unsigned long long fdt_be64(const void *p) {
   unsigned int hi = fdt_be32(p);
   unsigned int lo  = fdt_be32((const char *)p + 4);
-  return ((unsigned long)hi << 32) | lo;
+  return ((unsigned long long)hi << 32) | lo;
 }
 
 void fdt_init(void *fdt_ptr);
